Reduce latency on some keystrokes due to ncurses internal timer
Set the delay ncurses waits to determine whether a keystroke
is part of an escape sequence or not to 100ms instead of the
default 1000ms. This delay is painful in some cases, like
using Alt-2 in urxvt, which happens to be treated as part of
an escape sequence by curses, thus waits a full second to see
if something else happens.

The delay can now be set via the -d commandline switch.

Based on a patch from Pierre Habouzit, originally reported
as debian bug #512544