Overhaul color support
Don't assume default color values (-1,-1) to be white text
on black background. Because there doesn't seem to be a way
to read out the default colors we treat it as an additional
color. However this means that there are more than
COLOR_PAIRS possible color combinations. Therefore we always
use an internal mapping from the requested color to a valid
color pair. The mapping operates in LRU manner which means
if there are not enough color pairs available we wrap around
and invalidate/redefine the least recently used color pair.

Note that if you want to have more than 255 colors simultaneously
in use (i.e. you don't want that the 256 requested color pair
overwrites the first one) you have to build ncurses with
--enable-ext-colors which will set SOVERSION of the library to 6.

Also colors above COLORS are now rejected / ignored. Make
sure to set $TERM accordingly inorder for curses to pick up
the correct number of available colors.

Signed-off-by: Marc Andre Tanner <mat@brain-dump.org>