vt: correctly handle multibyte characters crossing read(2) boundaries
When a multibyte unicode character crosses 2 read(2) calls
we kept the whole byte sequence in our receive buffer but at
the same time we advanced the parsing state. This caused
a mismatch because the call to mbrtowc (after the second read(2))
expected (based on the given parsing state) only the remaining
bytes of the already started multibyte character, but we passed
in the whole sequence again.

Since we always keep the whole multibyte sequence in our buffer
we don't need to preserve the parsing state across multiple
function calls.

Signed-off-by: Marc Andre Tanner <mat@brain-dump.org>