Support \e[?10{47,48,49}{h,l} sequences
These were introduced by xterm to switch between normal and
alternate screen buffers. According to the documentation

 \e[1047 should switch the buffers
 \e[1048 should save/restore the cursor
 \e[1049 should do both

However dvtm currently associates the cursor state with the
buffer therefore a buffer switch will always restore the previous
cursor state of that buffer.

This is not standard comformant but so far did not expose any
show stopper bugs?

Also switching to the alternate buffer should clear it.