Bug fix for MOD after a mouse action
After a mouse action, the key_index needs to be set back to zero. You
can reproduce the bug by performing a mouse action, then trying to
perform a dvtm keyboard command. You'll see that the MOD key is not
matched with a key binding, so will print to the current terminal
instead.