Redraw border with URGENT_ATTR as soon as possible
When a visible window other than the selected one receives a bell
('\a'), I expect its border to be redrawn immediately with the
corresponding attributes (URGENT_ATTR). Currently this does not
happen: the border will remain the same until draw_border() is
eventually called as a result of other events, such as layout
changing.