Compare full tag names, and assign them to not only visible windows
Since tags can be set to strings longer than the default one byte long
digits, we need to compare the whole strings and not only the first
character.

Assigning a tag should not depend on the visibility of a window, so we
loop over all the clients (as opposed to only the visible ones).