Extend tag command syntax to allow flexible tag manipulations
 tag win +foo -bar

will add the tag `foo` and remove the tag `bar` from `win`.
If a tag is neither prefixed with a plus or minus sign, the
existing set of tags are replaced with the one tag given.

The resulting tags must be non empty i.e. each window must be
assigned to at least one tag otherwise the command will be
ignored.